ゲストログインの際の名前変更の表示をなくす&名前の更新をコンポーネントにする
